+++ js/userSidebar/setSidebarSwiper.js
@@ -1,24 +1,10 @@
 import { sidebarAction } from "../library/swiper";
+import { getLocalStorageData } from "../localStorage/getLocalStorageData";
 import { makeDOMwithProperties } from "../utils/dom";
 import moveToDetail from "../utils/movetoProductDetail";
-import { getLocalStorageData } from "../localStorage/getLocalStorageData";
+import setSidebar from "./setSidebar";
 
-export function setSidebar(data, router) {
-  let sidebarData = getLocalStorageData("sidebarData");
-  // sidebarData에 이미 데이터가 있으면 배열에서 객체 삭제하고 배열 앞에 다시 삽입
-  sidebarData.forEach((item, index) => {
-    if (item.id === data.id) {
-      sidebarData.splice(index, 1);
-    }
-  });
-  // 최근 본 상품은 최대 10개까지 저장 가능
-  if (sidebarData.length === 10) sidebarData.pop();
-  sidebarData.unshift({ id: data.id, thumbnail: data.thumbnail });
-  localStorage.setItem("sidebarData", JSON.stringify(sidebarData));
-  setSidebarSwiper(router);
-}
-
-export function setSidebarSwiper(router) {
+export default function setSidebarSwiper(router) {
   const sidebarData = getLocalStorageData("sidebarData");
   const swiperWrapperDiv = document.querySelector(
     "#sidebar-area .swiper-wrapper"
